requirements.txt
openai
soundfile
numpy
discord.py
python-dotenv
pydub
